splitup map_layer_management
